@@ -7,7 +7,7 @@ readme = "README.md"
 authors = [
     {name = 'Jens Diemer', email = 'git@jensdiemer.de'}
 ]
-requires-python = ">=3.9"  # Stay with 3.9 until YunoHost used >=Debian 11 (Bullseye)
+requires-python = ">=3.11"
 dependencies = [
     "django-fmd>=0.4.1",  # https://gitlab.com/jedie/django-find-my-device
     #
@@ -33,8 +33,9 @@ dev = [
     "pyflakes",  # https://github.com/PyCQA/pyflakes
     "codespell",  # https://github.com/codespell-project/codespell
     "EditorConfig",  # https://github.com/editorconfig/editorconfig-core-py
-    "safety",  # https://github.com/pyupio/safety
+    "pip-audit",  # https://github.com/pypa/pip-audit
     "mypy",  # https://github.com/python/mypy
+    "typeguard",  # https://github.com/agronholm/typeguard/
     "twine",  # https://github.com/pypa/twine
 
     # https://github.com/akaihola/darker
@@ -43,11 +44,9 @@ dev = [
     # https://github.com/pygments/pygments
     "darker[flynt, isort, color]",
 
-    # indirect depencies added because of bug:
-    # https://github.com/pypa/pip/issues/9644 / https://github.com/jazzband/pip-tools/issues/1866
-    # to avoid errors like:
-    # In --require-hashes mode, all requirements must have their versions pinned with ==. These do not: ...
-    "tomli",  # Only needed for Python <3.11
+    # indirect depencies for Python <=3.11
+    # Work-a-round for: https://github.com/jazzband/pip-tools/issues/1326
+    "backports.tarfile",
 ]
 
 [project.urls]
@@ -72,6 +71,14 @@ include = ["django_fmd_ynh*"]
 
 [tool.setuptools.dynamic]
 version = {attr = "django_fmd_ynh.__version__"}
+
+
+[tool.cli_base.pip_audit]
+# https://github.com/jedie/cli-base-utilities/blob/main/docs/pip_audit.md
+requirements=["requirements.dev.txt"]
+strict=true
+require_hashes=true
+ignore-vuln=[]
 
 
 [tool.darker]
@@ -157,6 +164,9 @@ initial_revision = "1f3a70e"
 initial_date = 2024-05-21T21:22:39+02:00
 cookiecutter_template = "https://github.com/jedie/cookiecutter_templates/"
 cookiecutter_directory = "yunohost_django_package"
+applied_migrations = [
+    "2cdc1d8", # 2024-08-25T19:00:41+02:00
+]
 
 [manageprojects.cookiecutter_context.cookiecutter]
 project_name = "django-fmd"
